+++ suave/setup.py
@@ -25,8 +25,14 @@ setup(
     # tests_require=['pytest'],
     entry_points={
         'console_scripts': [
-            'pipeline_detection = ' +
-                'suave.pipeline_node:main',
+            'pipeline_detection = suave.pipeline_node:main',
+            'pipeline_detection_wv = suave_.pipeline_detection_wv:main',
+            'spiral_search = suave.spiral_search_lc:main',
+            'follow_pipeline = suave.follow_pipeline_lc:main',
+            'const_dist_mission_no_adapt = ' +
+                'suave.const_dist_mission_no_adapt:main',
+            'time_constrained_mission_no_adapt = ' +
+                'suave.time_constrained_mission_no_adapt:main',
         ],
     },
 )
