+++ suave/suave/const_dist_mission_no_adapt.py
@@ -5,7 +5,6 @@ import threading
 
 from datetime import datetime
 from rclpy.executors import MultiThreadedExecutor
-from suave_metacontrol.mission_planner import MissionPlanner
 from system_modes_msgs.srv import ChangeMode
